refactor: reorganize dirs and rename components for clarity
